refactor(policy-engine): move the supersession detector from apps/curator into the govern package layer
What: relocate detectSupersession/computeTitleSimilarity to
packages/policy-engine/src/supersession/, typed against a minimal
SupersessionMemorySource interface (structurally satisfied by the store's
MemoryRepository); apps/curator keeps a re-export shim so its public
@qmd-team-intent-kb/curator surface and every call site (curator.ts, api
promotion-service) are unchanged.

Why: Wave-2 C3 extends the govern-decision eval (packages/eval-surface) to
score dedup/contradiction/supersession decisions through the REAL pipeline
code, and the dependency-cruiser no-package-depends-on-app invariant forbids
a package importing apps/curator. Chose extraction-into-a-package over
duplicating the detector in the eval (a copy would silently drift from the
production decision) — the depcruise config itself prescribes extraction.

How verified: pnpm typecheck green; policy-engine suite 95/95 and curator
suite 279/279 pass (the curator supersession tests now exercise the
re-export path against a real in-memory store).
